Rename TwilioTracer to TwilioSingletons
