+++ instrumentation/twilio-6.6/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/twilio/TwilioAsyncInstrumentation.java
@@ -8,8 +8,8 @@ package io.opentelemetry.javaagent.instrumentation.twilio;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.extendsClass;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.hasClassesNamed;
 import static io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge.currentContext;
-import static io.opentelemetry.javaagent.instrumentation.twilio.TwilioTracer.instrumenter;
-import static io.opentelemetry.javaagent.instrumentation.twilio.TwilioTracer.spanName;
+import static io.opentelemetry.javaagent.instrumentation.twilio.TwilioSingletons.instrumenter;
+import static io.opentelemetry.javaagent.instrumentation.twilio.TwilioSingletons.spanName;
 import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
